Post /Order

+++ backend/src/main/java/com/felipescherer/fsdeliver/dto/ProductDTO.java
@@ -8,7 +8,7 @@ public class ProductDTO implements Serializable{
 	
 	private static final long serialVersionUID = 1L;
 	
-	private Double id;
+	private Long id;
 	private String name;
 	private Double price;
 	private String description;
@@ -18,7 +18,7 @@ public class ProductDTO implements Serializable{
 		
 	}
 
-	public ProductDTO(Double id, String name, Double price, String description, String imageUri) {
+	public ProductDTO(Long id, String name, Double price, String description, String imageUri) {
 		this.id = id;
 		this.name = name;
 		this.price = price;
@@ -34,11 +34,11 @@ public class ProductDTO implements Serializable{
 		imageUri = entity.getImageUri();
 	}
 
-	public Double getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Double id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
